requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
apturl==0.5.2
async-timeout==5.0.1
attrs==21.2.0
bcrypt==3.2.0
bitarray==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
ckzg==2.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cytoolz==1.0.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fasteners==0.14.1
filelock==3.6.0
fonttools==4.55.8
frozenlist==1.5.0
future==0.18.2
gyp==0.1
hexbytes==1.3.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.10.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.2.0
netifaces==0.11.0
nose2==0.5.0
numpy==2.2.2
oauthlib==3.2.0
olefile==0.46
packaging==24.2
paramiko==2.9.3
parsimonious==0.10.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
propcache==0.3.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodome==3.22.0
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyunormalize==16.0.0
pyxdg==0.27
PyYAML==5.4.1
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
rlp==4.1.0
scipy==1.15.1
SecretStorage==3.3.1
six==1.10.0
ssh-import-id==5.11
systemd-python==234
texttable==1.6.4
toolz==1.0.0
types-requests==2.32.0.20250306
typing_extensions==4.13.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
web3==7.9.0
websocket-client==1.2.3
websockets==13.1
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
